+++ tests/Pest.php
@@ -3,23 +3,16 @@
 declare(strict_types=1);
 
 use CreativeCrafts\LaravelTwillioSms\Tests\TestCase;
+use Illuminate\Support\Facades\Config;
 
 uses(TestCase::class)->in(__DIR__);
 
 pest()->project()->github('CreativeCrafts/laravel-twillio-sms');
 
-
-/**
- * Helper function to access private properties in a class.
- *
- * @param object $object
- * @param string $propertyName
- * @return mixed
- */
-function getPrivateProperty($object, string $propertyName): mixed
+function setDefaultConfig(bool $riskCheck = true): void
 {
-    $reflection = new ReflectionClass($object);
-    $property = $reflection->getProperty($propertyName);
-    $property->setAccessible(true);
-    return $property->getValue($object);
+    Config::set('twillio-sms.risk_check', $riskCheck);
+    Config::set('twillio-sms.account_sid', 'valid-account-sid');
+    Config::set('twillio-sms.auth_token', 'valid-account-sid');
+    Config::set('twillio-sms.sms_from', '+1234567890');
 }